Moved dimwits quantities and R-matrix options to their own separate header with guards
