+++ src/resonanceReconstruction/rmatrix/options.hpp
@@ -1,0 +1,26 @@
+#ifndef NJOY_R2_RMATRIX_OPTIONS
+#define NJOY_R2_RMATRIX_OPTIONS
+
+// system includes
+
+// other includes
+
+namespace njoy {
+namespace resonanceReconstruction {
+namespace rmatrix {
+
+  // R-matrix boundary condition options
+  struct ShiftFactor {}; // shift factor eliminates the boundary condition
+  struct Constant {};    // the boundary condition is constannt
+
+  // R-matrix formalism options
+  struct SingleLevelBreitWigner {};
+  struct MultiLevelBreitWigner {};
+  struct ReichMoore {};
+  struct GeneralRMatrix {};
+
+} // rmatrix namespace
+} // resonanceReconstruction namespace
+} // njoy namespace
+
+#endif
